+++ lib/garner/cache/object_identity.rb
@@ -38,7 +38,8 @@ module Garner
         Garner::Strategies::Keys::Caller,
         Garner::Strategies::Keys::Version,
         Garner::Strategies::Keys::RequestPath,
-        Garner::Strategies::Keys::RequestGet
+        Garner::Strategies::Keys::RequestGet,
+        Garner::Strategies::Keys::RequestPost
       ]
 
       CACHE_STRATEGIES = [
@@ -164,7 +165,7 @@ module Garner
               find_or_create_key_prefix_for(el[:klass], el[:object])
             }.join(",") + ":" +
             Digest::MD5.hexdigest(
-              KEY_STRATEGIES.map { |strategy| context[strategy.field] }.compact.join("\n")
+              KEY_STRATEGIES.map { |strategy| context[strategy.field] }.uniq.compact.join("\n")
             )
           end